Add files via upload
This repository contains a C++ implementation of a function to determine if a binary tree is symmetric (i.e., a mirror of itself). The algorithm uses a recursive approach to compare the left and right subtrees of the tree.

Key Features:
Definition of the binary tree node structure (TreeNode).
Recursive function isSymmetric to check for symmetry.
Efficient time complexity of 
𝑂
(
𝑛
)
O(n), where 
𝑛
n is the number of nodes in the tree.