Add files via upload
Given an array of integers nums sorted in non-decreasing order, you are tasked with finding the starting and ending positions of a given target value. If the target is not found in the array, you should return [-1, -1].

The algorithm must be efficient, with a target time complexity of O(log N) due to the array being sorted. You are expected to use binary search to achieve this optimal time complexity.